pyproject.toml
[project]
name = "unichunk"
version = "0.1.0"
description = ""
authors = [
    {name = "Anand Ankit (MS/EMS3-BST)",email = "Ankit.Anand@in.bosch.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.0"
pylint = "^3.3.7"

